@@ -15,7 +15,7 @@
     "ext-zip": "*",
     "beberlei/doctrineextensions": "1.5.*",
     "bugsnag/bugsnag-symfony": "^1.14",
-    "catrobat/capi": "1.4.*",
+    "catrobat/capi": "1.5.*",
     "codercat/jwk-to-pem": "1.1.*",
     "doctrine/annotations": "2.0.*",
     "doctrine/dbal": "3.8.*",
@@ -41,27 +41,26 @@
     "sonata-project/doctrine-orm-admin-bundle": "4.17.*",
     "sonata-project/exporter": "3.3.*",
     "sonata-project/twig-extensions": "2.4.*",
-    "sonata-project/user-bundle": "5.12.*",
-    "symfony/acl-bundle": "2.3.*",
-    "symfony/asset": "6.4.*",
-    "symfony/brevo-mailer": "6.4.*",
-    "symfony/cache": "6.4.*",
-    "symfony/console": "6.4.*",
-    "symfony/dotenv": "6.4.*",
+    "symfony/asset": "7.0.*",
+    "symfony/brevo-mailer": "7.0.*",
+    "symfony/cache": "7.0.*",
+    "symfony/console": "7.0.*",
+    "symfony/dotenv": "7.0.*",
     "symfony/flex": "2.4.*",
-    "symfony/framework-bundle": "6.4.*",
-    "symfony/intl": "6.4.*",
-    "symfony/mailer": "6.4.*",
+    "symfony/framework-bundle": "7.0.*",
+    "symfony/intl": "7.0.*",
+    "symfony/mailer": "7.0.*",
     "symfony/monolog-bundle": "3.10.*",
     "symfony/polyfill-iconv": "1.28.*",
-    "symfony/process": "6.4.*",
-    "symfony/routing": "6.4.*",
+    "symfony/process": "7.0.*",
+    "symfony/routing": "7.0.*",
+    "symfony/runtime": "7.0.*",
     "symfony/stimulus-bundle": "2.13.*",
-    "symfony/translation": "6.4.*",
-    "symfony/twig-bundle": "6.4.*",
-    "symfony/validator": "6.4.*",
+    "symfony/translation": "7.0.*",
+    "symfony/twig-bundle": "7.0.*",
+    "symfony/validator": "7.0.*",
     "symfony/webpack-encore-bundle": "2.1.*",
-    "symfony/yaml": "6.4.*",
+    "symfony/yaml": "7.0.*",
     "symfonycasts/reset-password-bundle": "1.22.*",
     "symfonycasts/verify-email-bundle": "1.17.*",
     "twig/extra-bundle": "3.10.*",
@@ -72,7 +71,7 @@
     "behat/mink": "1.11.*",
     "deployer/deployer": "v7.4.*",
     "dmore/behat-chrome-extension": "1.4.*",
-    "dmore/chrome-mink-driver": "2.9.*",
+    "dmore/chrome-mink-driver": "2.9.2",
     "doctrine/doctrine-fixtures-bundle": "3.6.*",
     "dvdoug/behat-code-coverage": "5.3.*",
     "friends-of-behat/symfony-extension": "2.5.*",
@@ -85,11 +84,11 @@
     "phpunit/phpunit": "10.5.*",
     "rector/rector": "1.1.*",
     "symfony/browser-kit": "7.0.*",
-    "symfony/http-client": "6.4.*",
+    "symfony/http-client": "7.0.*",
     "symfony/maker-bundle": "1.58.*",
-    "symfony/phpunit-bridge": "6.4.*",
-    "symfony/stopwatch": "6.4.*",
-    "symfony/web-profiler-bundle": "6.4.*",
+    "symfony/phpunit-bridge": "7.0.*",
+    "symfony/stopwatch": "7.0.*",
+    "symfony/web-profiler-bundle": "7.0.*",
     "vimeo/psalm": "5.24.*",
     "wapmorgan/php-deprecation-detector": "2.0.*",
     "textalk/websocket": "1.5.8.*"
@@ -115,7 +114,8 @@
     "allow-plugins": {
       "composer/package-versions-deprecated": true,
       "symfony/flex": true,
-      "php-http/discovery": true
+      "php-http/discovery": true,
+      "symfony/runtime": true
     }
   },
   "autoload": {
@@ -147,7 +147,7 @@
   "extra": {
     "symfony": {
       "allow-contrib": true,
-      "require": "6.4.*"
+      "require": "7.0.*"
     }
   },
   "repositories": [
