@@ -13,7 +13,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     [
       'dbal' => [
         'url' => '%env(resolve:DATABASE_URL)%',
-        'server_version' => 'mariadb-10.3.11',
+        'server_version' => '10.11.7-MariaDB',
         'types' => ['json' => JsonType::class],
         'mapping_types' => ['enum' => 'string'],
       ],
