@@ -5,22 +5,22 @@ declare(strict_types=1);
 use Symfony\Component\Routing\Loader\Configurator\RoutingConfigurator;
 
 return static function (RoutingConfigurator $routingConfigurator): void {
-  $routingConfigurator->import('../../src/Application/Controller', 'annotation')
+  $routingConfigurator->import('../../src/Application/Controller', 'attribute')
     ->prefix('/{theme}/')
     ->requirements(['theme' => '%themeRoutes%'])
   ;
 
-  $routingConfigurator->import('../../src/System/Controller', 'annotation')
+  $routingConfigurator->import('../../src/System/Controller', 'attribute')
     ->prefix('/{theme}/')
     ->requirements(['theme' => 'system'])
   ;
 
-  $routingConfigurator->import('../../src/Admin/', 'annotation')
+  $routingConfigurator->import('../../src/Admin/', 'attribute')
     ->prefix('/admin/')
     ->requirements(['theme' => 'admin'])
   ;
 
-  $routingConfigurator->import('../../src/Api_deprecated/Controller', 'annotation')
+  $routingConfigurator->import('../../src/Api_deprecated/Controller', 'attribute')
     ->prefix('/{theme}/')
     ->requirements(['theme' => '%themeRoutes%'])
   ;
