@@ -1,6 +1,6 @@
 <?php
 
-require dirname(__DIR__).'/config/bootstrap.php';
+require_once dirname(__DIR__).'/vendor/autoload_runtime.php';
 
 use WebPConvert\WebPConvert;
 
