@@ -9,10 +9,10 @@ use App\Security\PasswordGenerator;
 use App\Security\TokenGenerator;
 use App\User\UserManager;
 use Google\Client;
-use Sonata\UserBundle\Model\UserInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Validator\ConstraintViolationListInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
