@@ -176,14 +176,14 @@ class ResetCommand extends Command
     $password = 'catroweb';
 
     CommandHelper::executeShellCommand(
-      ['bin/console', 'sonata:user:create', 'catroweb', 'catroweb@localhost.at', $password, '--super-admin'],
+      ['bin/console', 'catrobat:user:create', 'catroweb', 'catroweb@localhost.at', $password, '--super-admin'],
       ['timeout' => 300], 'Create default admin user named catroweb with password catroweb', $output
     );
     $counter = count($user_array);
 
     for ($i = 1; $i < $counter; ++$i) { // starting at one because of admin user
       CommandHelper::executeShellCommand(
-        ['bin/console', 'sonata:user:create', $user_array[$i], $user_array[$i].'@localhost.at', $password],
+        ['bin/console', 'catrobat:user:create', $user_array[$i], $user_array[$i].'@localhost.at', $password],
         ['timeout' => 300], 'Create default user named '.$user_array[$i].' with password catroweb', $output
       );
     }
